Change PackagingMaterialWeight Type (#25)

+++ src/EPR.PRN.ObligationCalculation.Application/DTOs/ApprovedSubmissionEntity.cs
@@ -8,5 +8,5 @@ public class ApprovedSubmissionEntity
     public Guid OrganisationId { get; set; }
     public string? SubmissionPeriod { get; set; }
     public string? PackagingMaterial { get; set; }
-    public double PackagingMaterialWeight { get; set; }
+    public int PackagingMaterialWeight { get; set; }
 }
